3 Things
Unfortunately, there were several bugs that I was unable to fix as I
neared the deadline.

In the incoming_v2, there are several changes. I hardcoded things for
the colors, but that ended up making all the enemies change to the same
colors+prevention of taking out enemies.

In the WhiteBox game, there are two colored lines: a enemy and the
player. Basically, it’s a timed matching game. It’s super simple, and
easily coded.