GameOver
It works— kind of. I rearranged the offset, though I’m not really sure
if that did anything. I added a game over, which occurs only if 5 enemy
lines are filled. There’s still the problem of all the enemy lines
changing to the same color, however it’s kind of playable. I’m not sure
how to make it winnable.